Update hyperlinks from Gerrit to rendered docs
This change replaces links to Gerrit changes with links to the
rendered HTML documentation on docs.openstack.org as well as
minor gramatical nitpicking.

I think we have a chicken and egg situation. The infra change to
enable docs generation [1] has been merged but the docs aren't
in the expected location [2] yet. I think this is because the
Jenkins job will only run when a change is merged and nothing
has been merged in the networking-sfc repo since the infra
change. I believe that merging this change should trigger the
Jenkins job which will create [2].

[1] https://review.openstack.org/#/c/221589/
[2] http://docs.openstack.org/developer/networking-sfc/portchain-ovs-driver-agent.html

Change-Id: I20fb684fc9c8932ec1ea083ff9077e668cc2c1cc