Support backward compatibility of L2population
This patch will resolve this reported bug.
https://bugs.launchpad.net/networking-sfc/+bug/1539493

Under the service chain across different network(subnet) scenario,
if there is no entry of next hop sf in local host, the packet will
be broadcasted to nexthop(sf), even if l2 pop is enabled. eg,
	host1(network1 vm#1)->host2(network2 vm#2).

So, below patch was raised to install a fdb entry to resolve this
issue.
	https://review.openstack.org/#/c/227285/ patch set 52
And this patch import the l2pop driver object. However, the L2population
driver was reconstructed in the mitaka, after upgrade to mitaka,
some error is accurred.

This patch will support backward compatibility with Liberty too.

Change-Id: I6fc8b84f2ffd3bd7129bfacef5f02ad5fa0d96bd