Expose classifier driver pre-commit error to the user
The precommit driver method should fail the creation of a flow classifier
if it does not match the specific driver limitations and conditions.
In this case, we should expose the drivers exception to the user, so he/she
can know what needs to be fixed.
The general error does not help the user.

Change-Id: I27e9e10548af65c4866a301c37ad8bcca436ef47