Change query for port-pair-group and port-chain id. Fix bug 1625186.
Bug: When two non admin users in two different tenants/projects create
port-chain-group or port-chain, the first user would succeed, but the
second user did later would get an internal error, and creation failed.

Cause: The error was rooted at the query to retrieve port-pair-groups and
port-chains in create function, which would only get groups or chains in the
tenant/project which the user belongs to. The id column is integer and
has to be unique, and go by sequence from 1. New id was calculated by
the total current rows of groups and chains plus 1. This would cause id lose
uniqueness, and new group and chain creation failed.

Fix: This change will make the query to retrieve all rows in the table
regardless of tenant/project.

Closes-Bug: 1625186

Change-Id: Ia7cd5a98ec7150b2a10a3dfe32782ae5a4c18c79