Remove last vestiges of oslo-incubator
This is a community goal for Ocata.  We're already not using any of
the code, but there are some other files and references left over.

These files used to be installed from oslo-incubator:
  tools/install_venv.py
  tools/install_venv_common.py
  tools/with_venv.sh
They were meant to be used by the legacy run_tests.sh, which does not
exist in this repo.

Change-Id: Ie31ff50d524f4b9fee3df9f57cf49ebd176ae34f
Partial-Bug: #1639103