Default chain parameters per parameter
Make each chain parameter have its own default value.

With this change, if the user creates a chain that specifies only the
'correlation' parameter, the 'symmetric' parameter will still be added,
and vice-versa. This allows the code to make more assumptions about
the definition of the chain, and the tests to be less specific about
the chains they want to simulate.

Change-Id: Idf1052e6bec92aad953e61542671316dff371b78