SfcOVSBridgeExt: rely on ovs_lib to use the right OF version
With this change, SfcOVSBridgeExt uses add_protocols instead of
deprecated set_protocols(), and relies on ovs_lib change
I16a35b5d6c54901899d24fc94bd3438c1f1be05e to ensure that run_ofctl
will use the right OF version.

The code is simplified, and the question of which version of
OF is configured, is addressed in the module defining the
SfcOVSBridgeExt bridge.

Change-Id: I25fb8f35f8236c59a7dca9685a47de0ae77846bc
Depends-On: I16a35b5d6c54901899d24fc94bd3438c1f1be05e