Correct dl_type for non-SFC-proxied hops
For OVS, unit tests have an incorrect protocol number associated to the
dl_type match in the ACROSS_SUBNET_TABLE, for the scenarios where
packets are supposed to go to non-SFC-proxied SFs. Consequently, the
code is incorrect and this results in a bug that makes MPLS chains not
work when deployed with non-SFC-proxied hops.

This patch corrects this.

It also makes the inclusion of the pp_corr flowrule field explicit where
it was missing. It defaulted to the correct value anyway, but making
it explicit makes the intent of the affected tests clearer.

It also adds constants for the IP and MPLS ethertypes,
to be used by the OVS ethertypes.

Change-Id: I14bd7745f09793002755c52c3f4b9f75c277c29a
Closes-Bug: #1681863