Fix UT failures with multiple flow classifiers
Only observed on CentOS, but could happen on other systems:
 AssertionError: Mismatched values: 'source_ip_prefix',
 expected: u'10.0.0.9', actual: u'10.0.0.10

The list of flow classifiers can be in a different order than the
creation one, which causes the mismatch.

These tests now sort (on a relevant key) flow classifiers and expected
values lists. The assert check was also rewritten from the deprecated
assertDictContainsSubset method, and works fine on python 2 and 3

Change-Id: I5d7a6b4902522926ec07d06c8a291c5f41adc5bf
Closes-Bug: #1666568