stop mocking vlantransparent LOG in UTs
Today the base tests module imports the neutron vlantransparent
extension to mock out its LOG. However when running the UTs without
the LOG mocked, I don't see any output pertaining to vlantransparent
extension.

This patch removes the import and mock of the vlantransparent extension
in an effort to decouple sub-projects from neutron. If there's a real
need to mock this out, we'll need to find a different way to make it
happen in the future to break such dependency.

Change-Id: I5b3d06c48670fe2f550ff96b331cadaae96ebe3b