Remove note about migration from lib/neutron-legacy to lib/neutron
In [1] we finally got rid of the unfinished lib/neutron module and kept
only lib/neutron-legacy. It's renamed to lib/neutron now and it's the
only neutron related module in Devstack.
So this patch removes left over todo comments about things to do when
migration to the new lib/neutron module will be finished.

[1] https://review.opendev.org/c/openstack/devstack/+/865014

Change-Id: I54ff765d1f839232c85ab01556be03d95f135bd5