Fix _delete_ports_flowrules_by_id so that it expects only one port_id
Private method _delete_ports_flowrules_by_id was done in the way that it
was expecting list of ports given as "ports_id" variable but the only
place where this method was called is passing ID of the one port only.
As it was not list but string, iteraton was done over string elements
and that was causing plenty of messages send to the neutron server for
every char of the each deleted port if. This was overloading
neutron-ovs-agent.

This patch changes that private method to be
_delete_port_flowrules_by_id and to expect only one port id to be passed
to it as argument as there is no any place in the code where old method
would be called with list of the ports.

Depends-On: https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/926915

Closes-bug: #2068727
Change-Id: I76099a4cae2b47a748478525c25882538390bb91